Task 6: Implement another page replacement policy
I think I implemented the SPN Policy? I'm not entirely sure if it works correctly. I probably didn't add some of the stuff I needed to in order to get the code to actually use the policy... I didn't know how to do the system test parts of the assignment cause I read it and did't know what it meant... so this is as far as I got. The code compiles.